refactor : 토큰 관련 코드 계층화(5)
